Use new Vercel project structure
Uses the new Vercel project structure.
