+++ README.org
@@ -13,13 +13,6 @@ API used to dynamically generate SVG and PNG Ludum Dare results badges in real-t
 *[[https://badges.vandoorn.ca][Live Demo]]*
 
 *Note:* This project is community made and not endorsed by the offical Ludum Dare creator(s).
-* Packages
-
-This monorepo is split up into the following packages:
-
-- *[[https://github.com/woofers/ludum-dare-badges/tree/master/packages/api][@ludum-dare-badges/api]]* - Serveless API used to generate SVG & PNG badges
-- *[[https://github.com/woofers/ludum-dare-badges/tree/master/packages/demo][@ludum-dare-badges/demo]]* - Demo site used to showcase how to build a API query for a badge
-- *[[https://github.com/woofers/ludum-dare-badges/tree/master/packages/react][@ludum-dare-badges/react]]* - React wrapper component which forms and calls the API based off passed props.
 
 * Examples
 
@@ -51,7 +44,7 @@ You can also think of it as adding ~/badge.svg~ to the existing Ludum Dare path:
 
 #+CAPTION: URL Path
 #+NAME:    URL Path
-[[./packages/api/screenshots/url.png]]
+[[./screenshots/url.png]]
 
 ** Hosted
 This path can be appended to the domain of ~badges.vandoorn.ca~ to form the image URL:
@@ -69,24 +62,24 @@ badges.vandoorn.ca/:id/:game/badge.<type>
 ~badges.vandoorn.ca~ may also be used as the domain.
 ** Self Hosting
 
-Self hosting can done easily though [[https://zeit.co/now][Now]]'s serverless platform.
+Self hosting can done easily though [[https://vercel.com/home][Vercel]]'s serverless platform.
 
-#+CAPTION: Deploy to now
-[[https://zeit.co/new/project?template=woofers/ludum-dare-badges][https://deploy.now.sh/static/button.svg]]
+#+CAPTION: Deploy with Vercel
+https://vercel.com/button
 
-Simply install the [[https://github.com/zeit/now-cli][Now CLI]] with:
+Simply install the [[https://github.com/vercel/vercel/tree/master/packages/now-cli][Vercel CLI]] with:
 
 *Yarn*
 #+BEGIN_SRC
-yarn global add now
+yarn global add vercel
 #+END_SRC
 
 *npm*
 #+BEGIN_SRC
-npm install -g now
+npm install -g vercel
 #+END_SRC
 
-and type ~now deploy~ in the root of the monorepo.
+and type ~vercel~ in the root of the monorepo.
 
 To form the API query follow the steps above with and replace the domain
 with the domain of your self hosted instance:
