+++ _react/package.json
@@ -23,7 +23,7 @@
   "bugs": {
     "url": "https://github.com/woofers/ludum-dare-badges/issues"
   },
-  "homepage": "https://github.com/woofers/ludum-dare-badges/tree/master/packages/react#readme",
+  "homepage": "https://github.com/woofers/ludum-dare-badges/tree/master/_react#readme",
   "keywords": [
     "react",
     "reactjs",
@@ -35,20 +35,20 @@
     "react": "^0.13.0 || ^0.14.0 || ^15.0.0 || ^16.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.5",
-    "@babel/plugin-proposal-class-properties": "^7.4.4",
-    "@babel/preset-env": "^7.4.5",
-    "@babel/preset-react": "^7.0.0",
-    "babel-eslint": "^10.0.2",
-    "babel-loader": "^8.0.6",
-    "eslint": "^6.1.0",
-    "eslint-loader": "^2.2.1",
-    "eslint-plugin-react-hooks": "^1.6.1",
+    "@babel/core": "^7.11.6",
+    "@babel/plugin-proposal-class-properties": "^7.10.4",
+    "@babel/preset-env": "^7.11.5",
+    "@babel/preset-react": "^7.10.4",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.1.0",
+    "eslint": "^7.8.1",
+    "eslint-loader": "^4.0.2",
+    "eslint-plugin-react-hooks": "^4.1.0",
     "peer-deps-externals-webpack-plugin": "^1.0.4",
-    "rimraf": "^3.0.0",
-    "webpack": "^4.39.2",
-    "webpack-cli": "^3.3.5",
-    "webpack-node-externals": "^1.7.2"
+    "rimraf": "^3.0.2",
+    "webpack": "^4.44.1",
+    "webpack-cli": "^3.3.12",
+    "webpack-node-externals": "^2.5.2"
   },
   "dependencies": {}
 }
