+++ api/_lib/util/truncate.js
