--- now.json
@@ -1,19 +1,0 @@
-{
-  "version": 2,
-  "name": "ludum-dare-badges",
-  "builds": [
-    { "src": "package.json", "use": "@now/static-build", "config": { "distDir": "packages/demo/public" } },
-    { "src": "packages/api/src/routes/*.js", "use": "@now/node" }
-  ],
-  "routes": [
-    { "src": "/badge.svg", "dest": "packages/api/src/routes/large.js" },
-    { "src": "/badge.png", "dest": "packages/api/src/routes/large.js?rasterize=true" },
-    { "src": "/(.*)/(.*)/badge.svg", "dest": "packages/api/src/routes/large.js?id=$1&game=$2" },
-    { "src": "/(.*)/(.*)/badge.png", "dest": "packages/api/src/routes/large.js?id=$1&game=$2&rasterize=true" },
-    { "handle": "filesystem" },
-    { "src": "/.*", "status": 404, "dest": "404" }
-  ],
-  "github": {
-    "silent": true
-  }
-}
