--- packages/api/README.org
@@ -1,89 +1,0 @@
-
-#+TITLE:    @ludum-dare-badges/api
-#+AUTHOR:	Jaxson Van Doorn
-#+EMAIL:	jaxson.vandoorn@gmail.com
-#+OPTIONS:  num:nil toc:nil
-
-* @ludum-dare-badges/api
-
-[[https://github.com/woofers/ludum-dare-badges/actions][https://github.com/woofers/ludum-dare-badges/workflows/build/badge.svg]]
-
-API used to dynamically generate SVG and PNG Ludum Dare results badges in real-time.
-
-*[[https://badges.vandoorn.ca][Live Demo]]*
-
-*Note:* This project is community made and not endorsed by the offical Ludum Dare creator(s).
-
-* Examples
-
-*SVG Badge:*
-
-#+CAPTION: SVG Badge
-#+NAME:    SVG Badge
-https://badges.vandoorn.ca/44/alien-e-x-p-a-n-s-i-o-n/badge.svg
-
-*PNG Badge:*
-
-#+CAPTION: PNG Badge
-#+NAME:    PNG Badge
-https://badges.vandoorn.ca/44/alien-e-x-p-a-n-s-i-o-n/badge.png
-
-* Usage
-Simply form the API query using your game's Ludum Dare entry page path.
-
-The path is in the form ~:event/:game/badge.<type>~ where:
-
-- ~event~ - Ludum Dare event number (eg. ~44~ for Ludum Dare 44)
-- ~game~ - Game title in the URL of the Ludum Dare entry page.
-- ~type~ - File type.  Can be either ~svg~ or ~png~.  ~svg~ is recommenced if it is supported.
-
-For example the SVG badge for [[https://ldjam.com/events/ludum-dare/44/alien-e-x-p-a-n-s-i-o-n][ALIEN, e x p a n s i o n .]]
-the path would be ~44/alien-e-x-p-a-n-s-i-o-n/badge.svg~
-
-You can also think of it as adding ~/badge.svg~ to the existing Ludum Dare path:
-
-#+CAPTION: URL Path
-#+NAME:    URL Path
-[[./screenshots/url.png]]
-
-** Hosted
-This path can be appended to the domain of ~badges.vandoorn.ca~ to form the image URL:
-
-#+BEGIN_SRC
-badges.vandoorn.ca/44/alien-e-x-p-a-n-s-i-o-n/badge.svg
-#+END_SRC
-
-Or more generalized:
-
-#+BEGIN_SRC
-badges.vandoorn.ca/:id/:game/badge.<type>
-#+END_SRC
-
-~badges.vandoorn.ca~ may also be used as the domain.
-** Self Hosting
-
-Self hosting can done easily though [[https://zeit.co/now][Now]]'s serverless platform.
-
-#+CAPTION: Deploy to now
-[[https://zeit.co/new/project?template=woofers/ludum-dare-badges][https://deploy.now.sh/static/button.svg]]
-
-Simply install the [[https://github.com/zeit/now-cli][Now CLI]] with:
-
-*Yarn*
-#+BEGIN_SRC
-yarn global add now
-#+END_SRC
-
-*npm*
-#+BEGIN_SRC
-npm install -g now
-#+END_SRC
-
-and type ~now deploy~ in the root of the monorepo.
-
-To form the API query follow the steps above with and replace the domain
-with the domain of your self hosted instance:
-
-#+BEGIN_SRC
-<example.com>/:id/:game/badge.<type>
-#+END_SRC
