--- packages/api/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@ludum-dare-badges/api",
-  "version": "1.1.0",
-  "description": "SVG & PNG badges for Ludum Dare Game Jam results",
-  "files": [
-    "src",
-    "License.txt",
-    "package.json"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/woofers/ludum-dare-badges.git"
-  },
-  "bugs": {
-    "url": "https://github.com/woofers/ludum-dare-badges/issues"
-  },
-  "homepage": "https://github.com/woofers/ludum-dare-badges/tree/master/packages/api#readme",
-  "author": "Jaxson Van Doorn <jaxson.vandoorn@gmail.com>",
-  "license": "MIT",
-  "dependencies": {
-    "axios": "^0.19.1",
-    "sanitize-html": "^1.20.1"
-  },
-  "devDependencies": {}
-}
