--- packages/api/src/routes/index.js
@@ -1,1 +1,0 @@
-export default (req, res) => res.send('Read the docs ... at some point in time')
