--- packages/demo/README.md
@@ -1,36 +1,0 @@
-
-
-# @ludum-dare-badges/demo
-
-[![img](https://github.com/woofers/ludum-dare-badges/workflows/build/badge.svg)](https://github.com/woofers/ludum-dare-badges/actions)
-
-Demo site used to showcase and build a **[@ludum-dare-badges/api](https://github.com/woofers/ludum-dare-badges/tree/master/packages/api)** query
-for a Ludum Dare results badge.
-
-Site is [Gatsby](https://www.gatsbyjs.org/) based and built using [Material UI](https://material-ui.com/).
-
-**[Live Demo](https://badges.vandoorn.ca)**
-
-**Note:** This project is community made and not endorsed by the offical Ludum Dare creator(s).
-
-
-# Usage
-
-
-### Running Locally
-
-Running locally for development can be done with:
-
--   `yarn install` or `npm install` to install the required modules.
--   `yarn dev` or `npm dev` to host locally in development mode.
-
-Now going to [localhost:8000](http://localhost:8000) will display the site.
-
-
-## Building for Deployment
-
-Building the static site can be done with:
-
--   `yarn build` or `npm build`.
-
-A `public` folder will be generated with the created bundle.
