--- packages/demo/README.org
@@ -1,31 +1,0 @@
-
-#+TITLE:    @ludum-dare-badges/demo
-#+AUTHOR:	Jaxson Van Doorn
-#+EMAIL:	jaxson.vandoorn@gmail.com
-#+OPTIONS:  num:nil toc:nil
-
-* @ludum-dare-badges/demo
-
-[[https://github.com/woofers/ludum-dare-badges/actions][https://github.com/woofers/ludum-dare-badges/workflows/build/badge.svg]]
-
-Demo site used to showcase and build a *[[https://github.com/woofers/ludum-dare-badges/tree/master/packages/api][@ludum-dare-badges/api]]* query
-for a Ludum Dare results badge.
-
-Site is [[https://www.gatsbyjs.org/][Gatsby]] based and built using [[https://material-ui.com/][Material UI]].
-
-*[[https://badges.vandoorn.ca][Live Demo]]*
-
-*Note:* This project is community made and not endorsed by the offical Ludum Dare creator(s).
-
-* Usage
-*** Running Locally
-Running locally for development can be done with:
-- ~yarn install~ or ~npm install~ to install the required modules.
-- ~yarn dev~ or ~npm dev~ to host locally in development mode.
-Now going to [[http://localhost:8000][localhost:8000]] will display the site.
-
-** Building for Deployment
-
-Building the static site can be done with:
-- ~yarn build~ or ~npm build~.
-A ~public~ folder will be generated with the created bundle.
