--- packages/demo/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@ludum-dare-badges/demo",
-  "version": "1.1.0",
-  "description": "SVG & PNG badges for Ludum Dare Game Jam results",
-  "author": "Jaxson Van Doorn <jaxson.vandoorn@gmail.com>",
-  "dependencies": {
-    "@emotion/core": "^10.0.22",
-    "@ludum-dare-badges/react": "../react",
-    "@material-ui/core": "^4.5.1",
-    "emotion-server": "^10.0.17",
-    "emotion-theming": "^10.0.19",
-    "gatsby": "^2.17.4",
-    "gatsby-plugin-emotion": "^4.1.12",
-    "gatsby-plugin-manifest": "^2.2.23",
-    "gatsby-plugin-material-ui": "^2.1.6",
-    "gatsby-plugin-react-helmet": "^3.1.13",
-    "gatsby-plugin-typography": "^2.3.14",
-    "react": "^16.9.0",
-    "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "^16.11.0",
-    "react-emotion": "^10.0.0",
-    "react-helmet": "^5.2.1",
-    "react-typography": "^0.16.19",
-    "typography": "^0.16.19",
-    "typography-plugin-code": "^0.16.19",
-    "typography-theme-lawton": "^0.16.19"
-  },
-  "license": "MIT",
-  "scripts": {
-    "build": "gatsby build && cp -r public ../../.",
-    "dev": "gatsby develop",
-    "start": "gatsby develop",
-    "local": "yarn add ../react ",
-    "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\"",
-    "clean": "rimraf public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/woofers/ludum-dare-badges.git"
-  },
-  "bugs": {
-    "url": "https://github.com/woofers/ludum-dare-badges/issues"
-  },
-  "homepage": "https://github.com/woofers/ludum-dare-badges/tree/master/packages/demo#readme",
-  "devDependencies": {
-    "rimraf": "^3.0.0"
-  }
-}
