+++ src/pages/404.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'gatsby'
 import Layout from '../components/layout'
-import Helmet from 'react-helmet'
+import { Helmet } from 'react-helmet'
 
 const NotFoundPage = () => (
   <Layout>
