+++ src/pages/index.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react'
-import Helmet from 'react-helmet'
+import { Helmet } from 'react-helmet'
 import Layout from '../components/layout'
 import Input from '../components/input'
 import Form from '../components/form'
